Refactor tests of `discoverProjectAddons` to use a project.
Previously, these tests were using an addon-like mock object
with `discovery.discoverProjectAddons`. This is already bizarre
(projects are quite different from addons), but with the changes
added in this PR to add more `init` instrumentation we now require
a `Project` in `discoverProjectAddons` (we call `project.name()`
which is a method on `project`, but a property on addons).

This refactors the test to properly use a project instead, and adds
a few more missing pieces to `MockCLI` object.